Update main, Seperate embedding function
